Implement page summary template.
Implemented the page summary template. This page will contain the results and general statistics regarding the election that was ran.

Most notably, it includes a chart which displays how the budget was allocated across the selected projects. Further, a summary table which displays what project got selected in each round, along with a chart that demonstrates how much funding the supporters of the project had in comparison to how much they could've had had they did not spent on previously selected projects. Finally, the user is able to click on a row to expand and reveal details regarding the round and project. For example, a hyperlink which will open the round analysis page on the specific round the user expanded the row on and sample explanations which describes why a certain project got selected/rejected as well as the funding lost to all other projects.

Modified visualisation.py in order to retrieve the variables necessary for the page summary and adjusted test file accordingly to avoid errors.